+++ src/bask_task.h
@@ -33,8 +33,8 @@ struct bask_task
 	struct bask_task* next;
 };
 
-int task_check_input_nbrs (int, short, short, short);
-int task_check_input (bask_core*, char*, char*, char*, char*, char*, short);
+short task_check_input_nbrs (int, short, short, short);
+short task_check_input (bask_core*, char*, char*, char*, char*, char*, short);
 short task_get_priority (char*);
 
 void task_free_ll (struct bask_task**);
